// developer-portal-ui/src/models/certificateResponse.ts
/**
 * Certificate Generator
 * Certificate Generator for Testing Purposes of PSD2 Sandbox Environment
 *
 * OpenAPI spec version: 1.0.0
 * Contact: swi@adorsys.de
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Certificate Response
 */
export interface CertificateResponse {
  algorithm?: string;
  encodedCert?: string;
  keyId?: string;
  privateKey?: string;
}